Refactored log2timeline front-end to tool #160
