+++ tests/cli/extraction_tool.py
@@ -15,17 +15,23 @@ from tests.cli import test_lib
 class ExtractionToolTest(test_lib.CLIToolTestCase):
   """Tests for the extraction tool object."""
 
+  # pylint: disable=protected-access
+
   _EXPECTED_OUTPUT_EXTRACTION_OPTIONS = u'\n'.join([
-      (u'usage: extraction_tool_test.py [--hashers HASHER_LIST]'
-       u' [--yara_rules PATH]'),
-      (u'                               [--parsers PARSER_LIST]'
-       u' [--preferred_year YEAR]'),
-      u'                               [-p] [--process_archives]',
-      u'                               [--skip_compressed_streams]',
+      u'usage: extraction_tool_test.py [--artifact_definitions PATH]',
+      (u'                               [--hashers HASHER_LIST] '
+       u'[--parsers PARSER_LIST]'),
+      u'                               [--preferred_year YEAR] [-p]',
+      u'                               [--process_archives]',
+      (u'                               [--skip_compressed_streams] '
+       u'[--yara_rules PATH]'),
       u'',
       u'Test argument parser.',
       u'',
       u'optional arguments:',
+      u'  --artifact_definitions PATH, --artifact-definitions PATH',
+      (u'                        Path to a directory containing artifact '
+       u'definitions.'),
       u'  --hashers HASHER_LIST',
       (u'                        Define a list of hashers to use by the tool. '
        u'This is a'),
@@ -107,6 +113,62 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
       u'                        (defaults to 125000)',
       u''])
 
+  def testGetParserPresetsInformation(self):
+    """Tests the _GetParserPresetsInformation function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    parser_presets_information = test_tool._GetParserPresetsInformation()
+    self.assertGreaterEqual(len(parser_presets_information), 1)
+
+    available_parser_names = [name for name, _ in parser_presets_information]
+    self.assertIn(u'linux', available_parser_names)
+
+  @shared_test_lib.skipUnlessHasTestFile([u'artifacts'])
+  def testParseArtifactDefinitionsOption(self):
+    """Tests the _ParseArtifactDefinitionsOption function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    options = test_lib.TestOptions()
+
+    options.artifact_definitions_path = self._GetTestFilePath([u'artifacts'])
+
+    test_tool._ParseArtifactDefinitionsOption(options)
+
+  def testParseExtractionOptions(self):
+    """Tests the _ParseExtractionOptions function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    options = test_lib.TestOptions()
+
+    test_tool._ParseExtractionOptions(options)
+
+  def testParsePerformanceOptions(self):
+    """Tests the _ParsePerformanceOptions function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    options = test_lib.TestOptions()
+
+    test_tool._ParsePerformanceOptions(options)
+
+  def testParseStorageOptions(self):
+    """Tests the _ParseStorageOptions function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    options = test_lib.TestOptions()
+
+    test_tool._ParseStorageOptions(options)
+
+  @shared_test_lib.skipUnlessHasTestFile([u'yara.rules'])
+  def testParseYaraRulesOption(self):
+    """Tests the _ParseYaraRulesOption function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    options = test_lib.TestOptions()
+
+    options.yara_rules_path = self._GetTestFilePath([u'yara.rules'])
+
+    test_tool._ParseYaraRulesOption(options)
+
   def testAddExtractionOptions(self):
     """Tests the AddExtractionOptions function."""
     argument_parser = argparse.ArgumentParser(
@@ -131,6 +193,83 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
     output = self._RunArgparseFormatHelp(argument_parser)
     self.assertEqual(output, self._EXPECTED_PERFORMANCE_OPTIONS)
 
+  def testListHashers(self):
+    """Tests the ListHashers function."""
+    output_writer = test_lib.TestOutputWriter(encoding=u'utf-8')
+    test_tool = extraction_tool.ExtractionTool(output_writer=output_writer)
+
+    test_tool.ListHashers()
+
+    output = output_writer.ReadOutput()
+
+    number_of_tables = 0
+    lines = []
+    for line in output.split(b'\n'):
+      line = line.strip()
+      lines.append(line)
+
+      if line.startswith(b'*****') and line.endswith(b'*****'):
+        number_of_tables += 1
+
+    self.assertIn(u'Hashers', lines[1])
+
+    lines = frozenset(lines)
+
+    self.assertEqual(number_of_tables, 1)
+
+    expected_line = b'md5 : Calculates an MD5 digest hash over input data.'
+    self.assertIn(expected_line, lines)
+
+  def testListParsersAndPlugins(self):
+    """Tests the ListParsersAndPlugins function."""
+    output_writer = test_lib.TestOutputWriter(encoding=u'utf-8')
+    test_tool = extraction_tool.ExtractionTool(output_writer=output_writer)
+
+    test_tool.ListParsersAndPlugins()
+
+    output = output_writer.ReadOutput()
+
+    number_of_tables = 0
+    lines = []
+    for line in output.split(b'\n'):
+      line = line.strip()
+      lines.append(line)
+
+      if line.startswith(b'*****') and line.endswith(b'*****'):
+        number_of_tables += 1
+
+    self.assertIn(u'Parsers', lines[1])
+
+    lines = frozenset(lines)
+
+    self.assertEqual(number_of_tables, 9)
+
+    expected_line = b'filestat : Parser for file system stat information.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = b'bencode_utorrent : Parser for uTorrent bencoded files.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = (
+        b'msie_webcache : Parser for MSIE WebCache ESE database files.')
+    self.assertIn(expected_line, lines)
+
+    expected_line = b'olecf_default : Parser for a generic OLECF item.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = b'plist_default : Parser for plist files.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = (
+        b'chrome_history : Parser for Chrome history SQLite database files.')
+    self.assertIn(expected_line, lines)
+
+    expected_line = b'ssh : Parser for SSH syslog entries.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = b'winreg_default : Parser for Registry data.'
+    self.assertIn(expected_line, lines)
+
   @shared_test_lib.skipUnlessHasTestFile([u'ímynd.dd'])
   def testParseOptions(self):
     """Tests the ParseOptions function."""
@@ -145,8 +284,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
     options.source = self._GetTestFilePath([u'ímynd.dd'])
 
     test_tool.ParseOptions(options)
-
-    # TODO: improve this test.
 
 
 if __name__ == '__main__':
